+++ dukan_presentation/src/commonMain/kotlin/net/thechance/mena/dukan/presentation/screen/manageDukan/content/ManageDukanProducts.kt
@@ -10,9 +10,9 @@ import mena.dukan_presentation.generated.resources.Res
 import mena.dukan_presentation.generated.resources.empty_shelf
 import mena.dukan_presentation.generated.resources.shelf_empty_body
 import mena.dukan_presentation.generated.resources.shelf_empty_title
-import net.thechance.mena.dukan.presentation.component.EmptyStateContent
-import net.thechance.mena.dukan.presentation.component.LoadingVerticalList
-import net.thechance.mena.dukan.presentation.component.productCard.LoadingProductCard
+import net.thechance.mena.dukan.presentation.component.loading.LoadingProductCard
+import net.thechance.mena.dukan.presentation.component.loading.LoadingVerticalList
+import net.thechance.mena.dukan.presentation.component.state.EmptyStateContent
 import net.thechance.mena.dukan.presentation.screen.manageDukan.component.ManageDukanProductsList
 import net.thechance.mena.dukan.presentation.util.pagination.Pager
 import net.thechance.mena.dukan.presentation.viewModel.manageDukan.ManageDukanUiState
